backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
sarvamai>=0.1.4
python-multipart==0.0.12
